requirements.txt
# requirements.txt
python-telegram-bot[ext]>=20.0
fastapi
uvicorn
requests
python-dotenv